merging latest from upstream

+++ git-tar/function/ops.go
@@ -273,9 +273,16 @@ func getRepositoryURL(e sdk.PushEvent, authToken tokener) (string, error) {
 	return cu, nil
 }
 
-func clone(pushEvent sdk.PushEvent) (string, error) {
+func clone(fetcher RepoFetcher, pushEvent sdk.PushEvent) (string, error) {
 	workDir := os.TempDir()
 	destPath := path.Join(workDir, path.Join(pushEvent.Repository.Owner.Login, pushEvent.Repository.Name))
+
+	if len(pushEvent.Repository.Owner.Login) == 0 {
+		return "", fmt.Errorf("login must be specified")
+	}
+	if len(pushEvent.Repository.Name) == 0 {
+		return "", fmt.Errorf("repo name must be specified")
+	}
 
 	if _, err := os.Stat(destPath); err == nil {
 		truncateErr := os.RemoveAll(destPath)
@@ -314,23 +321,8 @@ func clone(pushEvent sdk.PushEvent) (string, error) {
 		}
 	}
 
-	git := exec.Command("git", "clone", cloneURL)
-	git.Dir = path.Join(workDir, pushEvent.Repository.Owner.Login)
-	log.Println(git.Dir)
-	err = git.Start()
-	if err != nil {
-		return "", fmt.Errorf("Cannot start git: %t", err)
-	}
-
-	err = git.Wait()
-
-	git = exec.Command("git", "checkout", pushEvent.AfterCommitID)
-	git.Dir = destPath
-	err = git.Start()
-	if err != nil {
-		return "", fmt.Errorf("Cannot start git checkout: %t", err)
-	}
-	err = git.Wait()
+	fetcher.Clone(cloneURL, path.Join(workDir, pushEvent.Repository.Owner.Login))
+	fetcher.Checkout(pushEvent.AfterCommitID, destPath)
 
 	return destPath, err
 }
